+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 == Changelog ==
 
+= 1.0.1 =
+* Remove th/st from dates
+* Changed Post type on single sessions to Authors
+* Added link to single session from archive and single proceeding pages
+* Moved time|room to seperate line
+* Removed session availability info unless session requires registration
+
+
 = 1.0.0 =
 * Add Changelog